refactor(vehicles): shows vehicle table on profile page
