+++ src/app/profile/[id]/page.tsx
@@ -1,4 +1,4 @@
-import ProfileCard from "@/components/Profile/ProfileCard";
+import ProfileInfoSection from "@/sections/ProfileInfoSection";
 import { Skeleton } from "@nextui-org/react";
 import { Suspense } from "react";
 
@@ -8,16 +8,14 @@ export default async function Profile({
   params: { id: string };
 }) {
   return (
-    <section className="flex justify-center mt-10 p-4">
-      <Suspense
-        fallback={
-          <div className="w-full h-1/3 flex justify-center mt-10">
-            <Skeleton className="w-full h-full max-w-lg max-h-xl rounded-lg" />
-          </div>
-        }
-      >
-        <ProfileCard id={id} />
-      </Suspense>
-    </section>
+    <Suspense
+      fallback={
+        <div className="w-full h-1/3 flex justify-center mt-10">
+          <Skeleton className="w-full h-full max-w-lg max-h-xl rounded-lg" />
+        </div>
+      }
+    >
+      <ProfileInfoSection id={id} />
+    </Suspense>
   );
 }
